IconButtons are gray when there are Theme customizations available
